requirements.txt
torch>=2.5.1
torchvision>=0.20.1
transformers
numpy
opencv-python
pillow
matplotlib
tqdm
scikit-learn
scikit-image
tifffile
kagglehub
imagecodecs
fvcore
hydra-core>=1.2.0
omegaconf>=2.2.0
tensorboard
submitit
